pyproject.toml
[tool.poetry]
name = "tradingviewhook"
version = "0.1.0"
description = ""
authors = ["LJlkdskdjflsa <liliangjya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
django-ninja = "^0.19.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
